engine: extract esm/planets_3d.js
